# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Molecular Vision AR (MVAR) API Documentation",
        "description": "Dokumentasi RESTful API untuk Molecular Vision AR (MVAR) Backend. API ini digunakan oleh aplikasi mobile (Flutter / Unity) dan client lainnya.",
        "contact": {
            "name": "Tim Pengembang MVAR",
            "email": "dev@mvar.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Current API Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/profile": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Pembaruan Profil Pengguna",
                "description": "Memperbarui data profil pengguna yang sedang login (nama, email, sekolah, dan jenjang kelas).",
                "operationId": "74599296b1816be41ad4228975229966",
                "requestBody": {
                    "description": "Data profil yang akan diperbarui",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Budi Santoso",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "budi_baru@example.com",
                                        "maxLength": 255
                                    },
                                    "school": {
                                        "type": "string",
                                        "example": "SMAN 1 Jakarta",
                                        "nullable": true,
                                        "maxLength": 255
                                    },
                                    "id_degree": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil berhasil diperbarui.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profil berhasil diperbarui."
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id_user": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Budi Santoso"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "budi_baru@example.com"
                                                        },
                                                        "school": {
                                                            "type": "string",
                                                            "example": "SMAN 1 Jakarta",
                                                            "nullable": true
                                                        },
                                                        "id_degree": {
                                                            "type": "integer",
                                                            "example": 2,
                                                            "nullable": true
                                                        },
                                                        "id_role": {
                                                            "type": "integer",
                                                            "example": 2,
                                                            "nullable": true
                                                        },
                                                        "role": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 2
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "siswa (mobile)"
                                                                }
                                                            },
                                                            "type": "object",
                                                            "nullable": true
                                                        },
                                                        "degree": {
                                                            "properties": {
                                                                "id_degree": {
                                                                    "type": "integer",
                                                                    "example": 2
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Kelas 11"
                                                                }
                                                            },
                                                            "type": "object",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - Token tidak valid atau tidak disertakan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi form gagal.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email has already been taken."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/profile/password": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Pembaruan Kata Sandi (Password)",
                "description": "Memperbarui kata sandi akun pengguna dengan memverifikasi kata sandi lama terlebih dahulu.",
                "operationId": "40f40b3f02ad252300bd78e362f6ef95",
                "requestBody": {
                    "description": "Data kata sandi lama dan baru",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Rahasia123!"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PasswordBaru123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kata sandi berhasil diperbarui.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Kata sandi berhasil diperbarui."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - Token tidak valid atau tidak disertakan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Kata sandi lama tidak sesuai atau format password baru tidak valid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kata sandi lama tidak sesuai."
                                        },
                                        "errors": {
                                            "properties": {
                                                "current_password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Kata sandi lama tidak sesuai."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Gunakan token Sanctum dengan format: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}